Extract case insensitive name mapping test in MongoDB
