+++ plugin/trino-mongodb/src/test/java/io/trino/plugin/mongodb/TestMongoConnectorTest.java
@@ -51,7 +51,6 @@ import static io.trino.spi.connector.ConnectorMetadata.MODIFYING_ROWS_MESSAGE;
 import static io.trino.testing.TestingNames.randomNameSuffix;
 import static java.lang.String.format;
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static java.util.Locale.ENGLISH;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.testng.Assert.assertEquals;
@@ -549,69 +548,6 @@ public class TestMongoConnectorTest
     }
 
     @Test
-    public void testCaseInsensitive()
-            throws Exception
-    {
-        MongoCollection<Document> collection = client.getDatabase("testCase").getCollection("testInsensitive");
-        collection.insertOne(new Document(ImmutableMap.of("Name", "abc", "Value", 1)));
-
-        assertQuery("SHOW SCHEMAS IN mongodb LIKE 'testcase'", "SELECT 'testcase'");
-        assertQuery("SHOW TABLES IN testcase", "SELECT 'testinsensitive'");
-        assertQuery(
-                "SHOW COLUMNS FROM testcase.testInsensitive",
-                "VALUES ('name', 'varchar', '', ''), ('value', 'bigint', '', '')");
-
-        assertQuery("SELECT name, value FROM testcase.testinsensitive", "SELECT 'abc', 1");
-        assertUpdate("INSERT INTO testcase.testinsensitive VALUES('def', 2)", 1);
-
-        assertQuery("SELECT value FROM testcase.testinsensitive WHERE name = 'def'", "SELECT 2");
-        assertUpdate("DROP TABLE testcase.testinsensitive");
-    }
-
-    @Test
-    public void testCaseInsensitiveRenameTable()
-    {
-        MongoCollection<Document> collection = client.getDatabase("testCase_RenameTable").getCollection("testInsensitive_RenameTable");
-        collection.insertOne(new Document(ImmutableMap.of("value", 1)));
-        assertQuery("SHOW TABLES IN testcase_renametable", "SELECT 'testinsensitive_renametable'");
-        assertQuery("SELECT value FROM testcase_renametable.testinsensitive_renametable", "SELECT 1");
-
-        assertUpdate("ALTER TABLE testcase_renametable.testinsensitive_renametable RENAME TO testcase_renametable.testinsensitive_renamed_table");
-
-        assertQuery("SHOW TABLES IN testcase_renametable", "SELECT 'testinsensitive_renamed_table'");
-        assertQuery("SELECT value FROM testcase_renametable.testinsensitive_renamed_table", "SELECT 1");
-        assertUpdate("DROP TABLE testcase_renametable.testinsensitive_renamed_table");
-    }
-
-    @Test
-    public void testNonLowercaseViewName()
-    {
-        // Case insensitive schema name
-        MongoCollection<Document> collection = client.getDatabase("NonLowercaseSchema").getCollection("test_collection");
-        collection.insertOne(new Document(ImmutableMap.of("Name", "abc", "Value", 1)));
-
-        client.getDatabase("NonLowercaseSchema").createView("lowercase_view", "test_collection", ImmutableList.of());
-        assertQuery("SELECT value FROM nonlowercaseschema.lowercase_view WHERE name = 'abc'", "SELECT 1");
-
-        // Case insensitive view name
-        collection = client.getDatabase("test_database").getCollection("test_collection");
-        collection.insertOne(new Document(ImmutableMap.of("Name", "abc", "Value", 1)));
-
-        client.getDatabase("test_database").createView("NonLowercaseView", "test_collection", ImmutableList.of());
-        assertQuery("SELECT value FROM test_database.nonlowercaseview WHERE name = 'abc'", "SELECT 1");
-
-        // Case insensitive schema and view name
-        client.getDatabase("NonLowercaseSchema").createView("NonLowercaseView", "test_collection", ImmutableList.of());
-        assertQuery("SELECT value FROM nonlowercaseschema.nonlowercaseview WHERE name = 'abc'", "SELECT 1");
-
-        assertUpdate("DROP TABLE nonlowercaseschema.lowercase_view");
-        assertUpdate("DROP TABLE test_database.nonlowercaseview");
-        assertUpdate("DROP TABLE nonlowercaseschema.test_collection");
-        assertUpdate("DROP TABLE test_database.test_collection");
-        assertUpdate("DROP TABLE nonlowercaseschema.nonlowercaseview");
-    }
-
-    @Test
     public void testSelectView()
     {
         assertUpdate("CREATE TABLE test.view_base AS SELECT 'foo' _varchar", 1);
@@ -782,17 +718,6 @@ public class TestMongoConnectorTest
     {
         assertThat(query("SELECT name FROM TABLE(mongodb.system.query(database => 'tpch', collection => 'region', filter => '{}'))"))
                 .matches("SELECT name FROM region");
-    }
-
-    @Test
-    public void testNativeQueryWithCaseInSensitiveNameMatch()
-    {
-        String tableName = "Test_Case_Insensitive" + randomNameSuffix();
-        String schemaName = "Test_Case_Insensitive_Schema" + randomNameSuffix();
-        client.getDatabase(schemaName).getCollection(tableName).insertOne(new Document("field", "hello"));
-
-        assertThat(query("SELECT * FROM TABLE(mongodb.system.query(database => '" + schemaName.toLowerCase(ENGLISH) + "', collection => '" + tableName.toLowerCase(ENGLISH) + "', filter => '{}'))"))
-                .matches("VALUES CAST('hello' AS VARCHAR)");
     }
 
     @Test
